backend/requirements.txt
uvicorn==0.23.2
langserve[server]>=0.0.30
pydantic==2.7.4
tqdm==4.66.5
unstructured[all-docs]==0.15.1
langchain-core==0.2.27
langchain-experimental==0.0.64
python-dotenv==1.0.1
langchain-openai==0.1.20
langchain-community==0.2.11
tiktoken==0.7.0
psycopg==3.2.1
psycopg-binary==3.2.1
pinecone-client==5.0.1
grpcio-tools==1.65.4
langchain-pinecone==0.1.3
psycopg2-binary==2.9.9
asyncpg==0.29.0
poetry==1.6.1
fastapi==0.112.0
langserve==0.2.2
starlette==0.37.2


